+++ R/Db.R
@@ -1,23 +1,30 @@
-#' Retrieves the date(s) of birth in decimal format
+#' Date(s) of birth in decimal format
 #' 
-#' Retrieves the date(s) of birth from the database
+#' Retrieves date(s) of birth from the database
 #' 
 #' 
 #' @param idego vector of IDs of egos
-#' @param dLH Name of database. If dLH is missing, dataLH_F is used.
-#' @return Returns the dates of birth
-#' @author Frans Willekens
+#' @param d Name of database. If d is missing, dLH is used.
+#' @return Dates of birth (decimal format)
 #' @examples
+#' # Load the data
+#' data(dLH,package = "Families")
 #' 
 #' # Date of birth of first individual in database
-#' data(dataLH_F,package = "Families")
-#' Db(idego=1) 
+#' d <- dLH
+#' Db(idego=1,d=d) 
+#' 
 #' 
 #' @export Db
-Db <- function (idego,dLH)
-{if (missing(dLH)) 
-   {  dLH <- Families::dataLH_F
-   } 
-  dbd <- dLH$bdated[idego]
+Db <- function (idego,d=NULL)
+{
+  # Tests
+  test <- Tests(idego=idego,d=d)
+  idego <- test$idego
+  d <- test$d
+  
+  # Date of birth
+  dbd <- rep(NA,length(idego))
+  dbd[!is.na(idego)]  <- d$bdated[idego[!is.na(idego)]]
   return (dbd)
 }