+++ NEWS.md
@@ -1,3 +1,35 @@
-# Families 1.0.0
+Families 1.0.2 Published July 8 2022
 
-This is the first version. 
+# Families 2.0.0 Submitted March 2024
+
+# Changes in version 2.0.0
+
+## New functions
+
+-   Age()
+-   Alive()
+-   IDau()
+-   IDgch()
+-   IDsib()
+-   Kin\_long()
+-   Nkin()
+-   PlotAges()
+-   Tests()
+
+## Functions removed:
+
+-   eo(): the function is now part of VirtualPop
+-   Multistate()
+
+## Data
+
+-   Removed:
+    -   rates, which is available in VirtualPop
+    -   dpopus, which was used in the vignette
+-   Changed: dataLH\_F is replaced by dLH (USA 2021)
+
+## Vignette
+
+The vignette “Families and Kinship Networks in Virtual Populations. The
+‘Families’ package.” has been extended and is now a separate paper,
+which will hopefully be published in the coming months.